+++ metrics/src/config.rs
@@ -1,0 +1,19 @@
+use {
+    serde::Deserialize,
+    std::net::{IpAddr, Ipv4Addr, SocketAddr},
+};
+
+#[derive(Debug, Clone, Copy, Deserialize)]
+#[serde(deny_unknown_fields, default)]
+pub struct ConfigMetrics {
+    /// Endpoint of Prometheus service
+    pub endpoint: SocketAddr,
+}
+
+impl Default for ConfigMetrics {
+    fn default() -> Self {
+        Self {
+            endpoint: SocketAddr::new(IpAddr::V4(Ipv4Addr::new(127, 0, 0, 1)), 10123),
+        }
+    }
+}