+++ richat/src/metrics.rs
@@ -1,9 +1,9 @@
 use {
     crate::version::VERSION as VERSION_INFO,
-    metrics::{counter, describe_counter, describe_gauge},
+    ::metrics::{counter, describe_counter, describe_gauge},
     metrics_exporter_prometheus::{BuildError, PrometheusBuilder, PrometheusHandle},
     richat_filter::filter::FilteredUpdateType,
-    richat_shared::config::ConfigMetrics,
+    richat_metrics::ConfigMetrics,
     solana_sdk::clock::Slot,
     std::{borrow::Cow, future::Future},
     tokio::{
@@ -35,6 +35,7 @@ pub const PUBSUB_MESSAGES_SENT_COUNT_TOTAL: &str = "pubsub_messages_sent_count_t
 pub const PUBSUB_MESSAGES_SENT_BYTES_TOTAL: &str = "pubsub_messages_sent_bytes_total"; // x_subscription_id, subscription
 pub const RICHAT_CONNECTIONS_TOTAL: &str = "richat_connections_total"; // transport
 
+#[rustfmt::skip]
 pub fn setup() -> Result<PrometheusHandle, BuildError> {
     let handle = PrometheusBuilder::new().install_recorder()?;
 
@@ -53,61 +54,25 @@ pub fn setup() -> Result<PrometheusHandle, BuildError> {
 
     describe_counter!(BLOCK_MESSAGE_FAILED, "Block message reconstruction errors");
     describe_gauge!(CHANNEL_SLOT, "Latest slot in channel by commitment");
-    describe_gauge!(
-        CHANNEL_MESSAGES_TOTAL,
-        "Total number of messages in channel"
-    );
+    describe_gauge!(CHANNEL_MESSAGES_TOTAL, "Total number of messages in channel");
     describe_gauge!(CHANNEL_SLOTS_TOTAL, "Total number of slots in channel");
     describe_gauge!(CHANNEL_BYTES_TOTAL, "Total size of all messages in channel");
     describe_gauge!(GRPC_BLOCK_META_SLOT, "Latest slot in gRPC block meta");
-    describe_gauge!(
-        GRPC_BLOCK_META_QUEUE_SIZE,
-        "Number of gRPC requests to block meta data"
-    );
+    describe_gauge!(GRPC_BLOCK_META_QUEUE_SIZE, "Number of gRPC requests to block meta data");
     describe_counter!(GRPC_REQUESTS_TOTAL, "Number of gRPC requests per method");
     describe_gauge!(GRPC_SUBSCRIBE_TOTAL, "Number of gRPC subscriptions");
-    describe_counter!(
-        GRPC_SUBSCRIBE_MESSAGES_COUNT_TOTAL,
-        "Number of gRPC messages in subscriptions by type"
-    );
-    describe_counter!(
-        GRPC_SUBSCRIBE_MESSAGES_BYTES_TOTAL,
-        "Total size of gRPC messages in subscriptions by type"
-    );
-    describe_gauge!(
-        GRPC_SUBSCRIBE_CPU_SECONDS_TOTAL,
-        "CPU consumption of gRPC filters in subscriptions"
-    );
+    describe_counter!(GRPC_SUBSCRIBE_MESSAGES_COUNT_TOTAL, "Number of gRPC messages in subscriptions by type");
+    describe_counter!(GRPC_SUBSCRIBE_MESSAGES_BYTES_TOTAL, "Total size of gRPC messages in subscriptions by type");
+    describe_gauge!(GRPC_SUBSCRIBE_CPU_SECONDS_TOTAL, "CPU consumption of gRPC filters in subscriptions");
     describe_gauge!(PUBSUB_SLOT, "Latest slot handled in PubSub by commitment");
-    describe_gauge!(
-        PUBSUB_CACHED_SIGNATURES_TOTAL,
-        "Number of cached signatures"
-    );
-    describe_gauge!(
-        PUBSUB_STORED_MESSAGES_COUNT_TOTAL,
-        "Number of stored filtered messages in cache"
-    );
-    describe_gauge!(
-        PUBSUB_STORED_MESSAGES_BYTES_TOTAL,
-        "Total size of stored filtered messages in cache"
-    );
+    describe_gauge!(PUBSUB_CACHED_SIGNATURES_TOTAL, "Number of cached signatures");
+    describe_gauge!(PUBSUB_STORED_MESSAGES_COUNT_TOTAL, "Number of stored filtered messages in cache");
+    describe_gauge!(PUBSUB_STORED_MESSAGES_BYTES_TOTAL, "Total size of stored filtered messages in cache");
     describe_gauge!(PUBSUB_CONNECTIONS_TOTAL, "Number of connections to PubSub");
-    describe_gauge!(
-        PUBSUB_SUBSCRIPTIONS_TOTAL,
-        "Number of subscriptions by type"
-    );
-    describe_counter!(
-        PUBSUB_MESSAGES_SENT_COUNT_TOTAL,
-        "Number of sent filtered messages by type"
-    );
-    describe_counter!(
-        PUBSUB_MESSAGES_SENT_BYTES_TOTAL,
-        "Total size of sent filtered messages by type"
-    );
-    describe_gauge!(
-        RICHAT_CONNECTIONS_TOTAL,
-        "Total number of connections to Richat"
-    );
+    describe_gauge!(PUBSUB_SUBSCRIPTIONS_TOTAL, "Number of subscriptions by type");
+    describe_counter!(PUBSUB_MESSAGES_SENT_COUNT_TOTAL, "Number of sent filtered messages by type");
+    describe_counter!(PUBSUB_MESSAGES_SENT_BYTES_TOTAL, "Total size of sent filtered messages by type");
+    describe_gauge!(RICHAT_CONNECTIONS_TOTAL, "Total number of connections to Richat");
 
     Ok(handle)
 }
@@ -125,7 +90,7 @@ pub async fn spawn_server(
         }
     });
 
-    richat_shared::metrics::spawn_server(
+    richat_metrics::spawn_server(
         config,
         move || handle.render().into_bytes(), // metrics
         || true,                              // health